add default values based on sw package

+++ bin/10_setup_db.sh
@@ -22,42 +22,66 @@
 # Modified...:
 # see git revision history for more information on changes/updates
 # ---------------------------------------------------------------------------
-# - Environment Variables ---------------------------------------------------
+# - Customization -----------------------------------------------------------
+# DB_BASE_PKG="LINUX.X64_193000_db_home.zip"          # Oracle Database 19c (19.3)
+# DB_BASE2_PKG=""                                     # Second Oracle Database package used for Oracle < 12.2
+# DB_EXAMPLE_PKG="LINUX.X64_193000_examples.zip"      # Oracle Examples 19c (19.3)
+# DB_PATCH_PKG="p30557433_190000_Linux-x86-64.zip"    # DATABASE RELEASE UPDATE 19.6.0.0.0 (Patch)
+# DB_OJVM_PKG="p30484981_190000_Linux-x86-64.zip"     # OJVM RELEASE UPDATE 19.6.0.0.0 (Patch)
+# DB_OPATCH_PKG="p6880880_190000_Linux-x86-64.zip"    # OPatch 12.2.0.1.17 for DB 19.x releases (APR 2019)
+# ORACLE_HOME_NAME="19.0.0.0"                         # Name of the Oracle Home directory
+# ORACLE_HOME="${ORACLE_BASE}/product/${ORACLE_HOME_NAME}"
+# ORACLE_MAJOR_RELEASE="190"                          # Oracle Major Release
+# ORACLE_EDITION="EE"                                 # Oracle edition EE or SE2
+# RESPONSE_FILE_VERSION="oracle.install.responseFileVersion=/oracle/install/rspfmt_dbinstall_response_schema_v19.0.0"
+# - End of Customization ----------------------------------------------------
+
+# - Default Values ----------------------------------------------------------
 # source genric environment variables and functions
 source "$(dirname ${BASH_SOURCE[0]})/00_setup_oradba_init.sh"
 
-# define the software packages
-export DB_BASE_PKG=${DB_BASE_PKG:-"LINUX.X64_180000_db_home.zip"}
+# define the default software packages
+export DB_BASE_PKG=${DB_BASE_PKG:-"LINUX.X64_193000_db_home.zip"}
 export DB_EXAMPLE_PKG=${DB_EXAMPLE_PKG:-""}
 export DB_PATCH_PKG=${DB_PATCH_PKG:-""}
 export DB_OJVM_PKG=${DB_OJVM_PKG:-""}
-export DB_OPATCH_PKG=${DB_OPATCH_PKG:-"p6880880_180000_Linux-x86-64.zip"}
-export RESPONSFILE_VERSION=${RESPONSFILE_VERSION:-"oracle.install.responseFileVersion=/oracle/install/rspfmt_dbinstall_response_schema_v18.0.0"}
+export DB_OPATCH_PKG=${DB_OPATCH_PKG:-"p6880880_190000_Linux-x86-64.zip"}
+
+# get default major release based on DB_BASE_PKG
+DEFAULT_ORACLE_MAJOR_RELEASE=$(echo $DB_BASE_PKG|cut -d_ -f2|cut -c1-3)
+if [ $DEFAULT_ORACLE_MAJOR_RELEASE -gt 122 ]; then
+    DEFAULT_ORACLE_MAJOR_RELEASE=$(echo $DEFAULT_ORACLE_MAJOR_RELEASE|sed 's/.$/0/')
+fi
+export ORACLE_MAJOR_RELEASE=${ORACLE_MAJOR_RELEASE:-$DEFAULT_ORACLE_MAJOR_RELEASE}
+
+# Set default response file version based on major release
+export RESPONSE_FILE_VERSION=${RESPONSE_FILE_VERSION:-"oracle.install.responseFileVersion=/oracle/install/rspfmt_dbinstall_response_schema_v${ORACLE_MAJOR_RELEASE:0:2}.${ORACLE_MAJOR_RELEASE:2:2}.0"}
+export ORACLE_EDITION=${ORACLE_EDITION:-"EE"}   # Oracle edition EE or SE2
 
 # define oradba specific variables
 export ORADBA_BIN="$(cd "$(dirname "${BASH_SOURCE[0]}")" ; pwd -P)"
 export ORADBA_BASE="$(dirname ${ORADBA_BIN})"
 export ORADBA_RSP="${ORADBA_BASE}/rsp"          # oradba init response file folder
+export ORADBA_DEBUG=${ORADBA_DEBUG:-"FALSE"}    # enable debug mode
 
-# define Oracle specific variables
+# define default Oracle specific environment variables
 export ORACLE_ROOT=${ORACLE_ROOT:-"/u00"}       # root folder for ORACLE_BASE and binaries
 export ORACLE_DATA=${ORACLE_DATA:-"/u01"}       # Oracle data folder eg volume for docker
 export ORACLE_ARCH=${ORACLE_ARCH:-"/u02"}       # Oracle arch folder eg volume for docker
 export ORACLE_BASE=${ORACLE_BASE:-"${ORACLE_ROOT}/app/oracle"}
 export ORACLE_INVENTORY=${ORACLE_INVENTORY:-"${ORACLE_ROOT}/app/oraInventory"}
-export ORACLE_EDITION=${ORACLE_EDITION:-"EE"}   # Oracle edition EE or SE2
-export ORACLE_HOME_NAME=${ORACLE_HOME_NAME:-"18.3.0.0"}
+# Set default Oracle home name based on major release
+export ORACLE_HOME_NAME=${ORACLE_HOME_NAME:-"${ORACLE_MAJOR_RELEASE:0:2}.${ORACLE_MAJOR_RELEASE:2:2}.0.0"}
 export ORACLE_HOME="${ORACLE_HOME:-${ORACLE_BASE}/product/${ORACLE_HOME_NAME}}"
-export ORACLE_MAJOR_RELEASE=${ORACLE_MAJOR_RELEASE:-"180"}
 
-# define generic variables for software, download etc
+# define generic environment variables for software, download etc
 export OPT_DIR=${OPT_DIR:-"/opt"}
 export SOFTWARE=${SOFTWARE:-"${OPT_DIR}/stage"} # local software stage folder
 export SOFTWARE_REPO=${SOFTWARE_REPO:-""}       # URL to software for curl fallback
 export DOWNLOAD=${DOWNLOAD:-"/tmp/download"}    # temporary download location
 export CLEANUP=${CLEANUP:-"true"}               # Flag to set yum clean up
 export SLIM=${SLIM:-"false"}                    # flag to enable SLIM setup
-# - EOF Environment Variables -----------------------------------------------
+# - End of Default Values ---------------------------------------------------
 
 # - Initialization ----------------------------------------------------------
 # Make sure root does not run our script
@@ -67,23 +91,23 @@ if [ ! $EUID -ne 0 ]; then
 fi
 
 # show what we will create later on...
-echo " - Settings -----------------------------------------------------------" && \
-echo "ORACLE_ROOT           =${ORACLE_ROOT}" && \
-echo "ORACLE_DATA           =${ORACLE_DATA}" && \
-echo "ORACLE_ARCH           =${ORACLE_ARCH}" && \
-echo "ORACLE_BASE           =${ORACLE_BASE}" && \
-echo "ORACLE_HOME           =${ORACLE_HOME}" && \
-echo "ORACLE_INVENTORY      =${ORACLE_INVENTORY}" && \
-echo "ORACLE_EDITION        =${ORACLE_EDITION}" && \
-echo "ORACLE_MAJOR_RELEASE  =${ORACLE_MAJOR_RELEASE}" && \
-echo "SOFTWARE              =${SOFTWARE}" && \
-echo "DOWNLOAD              =${DOWNLOAD}" && \
-echo "DB_BASE_PKG           =${DB_BASE_PKG}" && \
-echo "DB_EXAMPLE_PKG        =${DB_EXAMPLE_PKG}" && \
-echo "DB_PATCH_PKG          =${DB_PATCH_PKG}" && \
-echo "DB_OJVM_PKG           =${DB_OJVM_PKG}" && \
-echo "DB_OPATCH_PKG         =${DB_OPATCH_PKG}" && \
-echo "RESPONSFILE_VERSION   =${RESPONSFILE_VERSION}"
+echo " - Prepare Oracle DB binaries installation ----------------------------"
+echo "ORACLE_ROOT           =${ORACLE_ROOT}"
+echo "ORACLE_DATA           =${ORACLE_DATA}"
+echo "ORACLE_ARCH           =${ORACLE_ARCH}"
+echo "ORACLE_BASE           =${ORACLE_BASE}"
+echo "ORACLE_HOME           =${ORACLE_HOME}"
+echo "ORACLE_INVENTORY      =${ORACLE_INVENTORY}"
+echo "ORACLE_EDITION        =${ORACLE_EDITION}"
+echo "ORACLE_MAJOR_RELEASE  =${ORACLE_MAJOR_RELEASE}"
+echo "SOFTWARE              =${SOFTWARE}"
+echo "DOWNLOAD              =${DOWNLOAD}"
+echo "DB_BASE_PKG           =${DB_BASE_PKG}"
+echo "DB_EXAMPLE_PKG        =${DB_EXAMPLE_PKG}"
+echo "DB_PATCH_PKG          =${DB_PATCH_PKG}"
+echo "DB_OJVM_PKG           =${DB_OJVM_PKG}"
+echo "DB_OPATCH_PKG         =${DB_OPATCH_PKG}"
+echo "RESPONSE_FILE_VERSION   =${RESPONSE_FILE_VERSION}"
 
 # check space
 echo " - Check available space ----------------------------------------------"
@@ -106,7 +130,7 @@ sed -i -e "s|###ORACLE_BASE###|$ORACLE_BASE|g"              /tmp/db_install.rsp
 sed -i -e "s|###ORACLE_HOME###|$ORACLE_HOME|g"              /tmp/db_install.rsp
 sed -i -e "s|###ORACLE_INVENTORY###|$ORACLE_INVENTORY|g"    /tmp/db_install.rsp
 sed -i -e "s|###ORACLE_EDITION###|$ORACLE_EDITION|g"        /tmp/db_install.rsp
-sed -i -e "s|^oracle.install.responseFileVersion.*|$RESPONSFILE_VERSION|" /tmp/db_install.rsp
+sed -i -e "s|^oracle.install.responseFileVersion.*|$RESPONSE_FILE_VERSION|" /tmp/db_install.rsp
 
 # adjust response file for 11.2 and 12.1
 if [ ${ORACLE_MAJOR_RELEASE} -eq 112 ]; then
@@ -123,20 +147,7 @@ cp ${ORADBA_RSP}/db_examples_install.rsp.tmpl /tmp/db_examples_install.rsp
 sed -i -e "s|###ORACLE_INVENTORY###|$ORACLE_INVENTORY|g"    /tmp/db_examples_install.rsp
 sed -i -e "s|###ORACLE_BASE###|$ORACLE_BASE|g"              /tmp/db_examples_install.rsp
 sed -i -e "s|###ORACLE_HOME###|$ORACLE_HOME|g"              /tmp/db_examples_install.rsp
-sed -i -e "s|^oracle.install.responseFileVersion.*|$RESPONSFILE_VERSION|" /tmp/db_examples_install.rsp
-
-# - Prepare database binaries installation ----------------------------------
-echo " - Prepare Oracle DB binaries -----------------------------------------"
-echo " DB_BASE_PKG              =${DB_BASE_PKG}"
-echo " DB_EXAMPLE_PKG           =${DB_EXAMPLE_PKG}"
-echo " DB_PATCH_PKG             =${DB_PATCH_PKG}"
-echo " DB_OJVM_PKG              =${DB_OJVM_PKG}"
-echo " DB_OPATCH_PKG            =${DB_OPATCH_PKG}"
-echo " ORACLE_BASE              =${ORACLE_BASE}"
-echo " ORACLE_HOME              =${ORACLE_HOME}"
-echo " ORACLE_EDITION           =${ORACLE_EDITION}"
-echo " RESPONSFILE_VERSION      =${RESPONSFILE_VERSION}"
-
+sed -i -e "s|^oracle.install.responseFileVersion.*|$RESPONSE_FILE_VERSION|" /tmp/db_examples_install.rsp
 
 # - Install database binaries -----------------------------------------------
 # handle pre and post 18c
@@ -224,31 +235,42 @@ fi
 
 echo " - CleanUp DB installation --------------------------------------------"
 # Remove not needed components
-rm -rf ${ORACLE_HOME}/apex                  # APEX
-rm -rf ${ORACLE_HOME}/ords                  # ORDS
-rm -rf ${ORACLE_HOME}/bin/oracle_*          # Oracle Binaries
-rm -rf ${ORACLE_HOME}/sqldeveloper          # SQL Developer
-rm -rf ${ORACLE_HOME}/inventory/backup/*    # OUI backup
-rm -rf ${ORACLE_HOME}/network/tools/help    # Network tools help
-rm -rf ${ORACLE_HOME}/assistants/dbua       # Database upgrade assistant
-rm -rf ${ORACLE_HOME}/dmu                   # Database migration assistant
-rm -rf ${ORACLE_HOME}/install/pilot         # Remove pilot workflow installer
-rm -rf ${ORACLE_HOME}/suptools              # Support tools
-rm -rf ${ORACLE_HOME}/ucp                   # UCP connection pool
-rm -rf ${ORACLE_HOME}/lib/*.zip             # All installer files
+if running_in_docker; then
+    echo "INFO:    remove Docker specific stuff"
+    rm -rf ${ORACLE_HOME}/apex                  # APEX
+    rm -rf ${ORACLE_HOME}/ords                  # ORDS
+    rm -rf ${ORACLE_HOME}/bin/oracle_*          # Oracle Binaries
+    rm -rf ${ORACLE_HOME}/sqldeveloper          # SQL Developer
+    rm -rf ${ORACLE_HOME}/inventory/backup/*    # OUI backup
+    rm -rf ${ORACLE_HOME}/network/tools/help    # Network tools help
+    rm -rf ${ORACLE_HOME}/assistants/dbua       # Database upgrade assistant
+    rm -rf ${ORACLE_HOME}/dmu                   # Database migration assistant
+    rm -rf ${ORACLE_HOME}/install/pilot         # Remove pilot workflow installer
+    rm -rf ${ORACLE_HOME}/suptools              # Support tools
+    rm -rf ${ORACLE_HOME}/ucp                   # UCP connection pool
+    rm -rf ${ORACLE_HOME}/lib/*.zip             # All installer files
+fi
 
-# Temp locations
-rm -rf ${DOWNLOAD}/*
-rm -rf /tmp/*.rsp
-rm -rf /tmp/InstallActions*
-rm -rf /tmp/CVU*oracle
-rm -rf /tmp/OraInstall*
-
-# remove all the logs....
-find ${ORACLE_INVENTORY} -type f -name *.log -exec rm {} \;
-find ${ORACLE_BASE}/product -type f -name *.log -exec rm {} \;
+if [ $ORADBA_DEBUG == ""]
+if [ "${ORADBA_DEBUG^^}" == "TRUE" ]; then
+    echo "INFO:    \$ORADBA_DEBUG set to TRUE, keep temp and log files"
+else
+    echo "INFO:    \$ORADBA_DEBUG not set, remove temp and log files"
+    # Temp locations
+    echo "INFO:    remove temp files"
+    rm -rf ${DOWNLOAD}/*
+    rm -rf /tmp/*.rsp
+    rm -rf /tmp/InstallActions*
+    rm -rf /tmp/CVU*oracle
+    rm -rf /tmp/OraInstall*
+    # remove all the logs....
+    echo "INFO:    remove log files in \${ORACLE_INVENTORY} and \${ORACLE_BASE}/product"
+    find ${ORACLE_INVENTORY} -type f -name *.log -exec rm {} \;
+    find ${ORACLE_BASE}/product -type f -name *.log -exec rm {} \;
+fi
 
 if [ "${SLIM^^}" == "TRUE" ]; then
+    echo "INFO:    \$SLIM set to TRUE, remove other stuff..."
     rm -rf ${ORACLE_HOME}/inventory             # remove inventory
     rm -rf ${ORACLE_HOME}/oui                   # remove oui
     rm -rf ${ORACLE_HOME}/OPatch                # remove OPatch
